added scaffolding and created real tests
The "scaffolding" refers to the data structure - as I thought I should at least define that to have a clear way of writing my tests. Then , I created three tests but only wrote one function.

After I have a better idea as to how robust my program is meant to be (ie should I make sure that the nodes of which we need to find the LCA exist within the binary tree, among other things) I will complete the tests and write the corresponding functions