added new test and then wrote corresponding function
I deleted the "test_exists" test because I realized that the function would be trivial - as it would be wholly dependent on another function which creates a list for which every element is a node of a binary tree. With this function, which I will call tree_list, I decided to create a simple test. I then wrote a simple function and it broke my code 1000 ways. The next step is to finish the test - make it as robust as I want it to be - and then finish up the code.