scrapping a whole function
The idea was to create a list which contains a Nodes path and use it to compare the nodes of two paths to then get their LCA. However, I realized that the idea was awkward, error-prone, and overly complex. I found an algorithm which is a lot simpler I don't need to change my tests beforee implementing it(obviously).