Finished the programme
As described in the last commit, I realized that most of my tests were necessary - I overcomplicated the problem. Thus,  I took out the functions which helped identify if a tree was valid (properly defined), and added some extra tests to the LCA function.

The result is only one test and one corresponding method, which, using the unit testing, works as I wanted it to.